Move regex/ast package to regex/parser/ast
