+++ internal/regex/parser/ast/ast_test.go
